# libs/contracts/openapi/auth.yaml
openapi: 3.1.0
info:
  title: Auth Service API
  version: 1.0.0
  description: |
    Authentication and identity service. Issues JWT access tokens (RS256) and
    refresh tokens. Publishes user lifecycle events to Kafka via the
    transactional outbox.
servers:
  - url: http://auth-service:8081
    description: Local docker network
paths:
  /healthz:
    get: { summary: Liveness, responses: { '200': { description: OK } } }
  /readyz:
    get: { summary: Readiness, responses: { '200': { description: Ready } } }
  /.well-known/jwks.json:
    get:
      summary: JSON Web Key Set (public keys for JWT verification)
      responses:
        '200':
          description: JWKS
          content: { application/json: { schema: { type: object } } }
  /api/v1/auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Created
          content: { application/json: { schema: { $ref: '#/components/schemas/UserResponse' } } }
        '409': { $ref: '#/components/responses/Conflict' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/v1/auth/login:
    post:
      summary: Username + password login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Token pair
          content: { application/json: { schema: { $ref: '#/components/schemas/TokenPair' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/refresh:
    post:
      summary: Exchange refresh token for a new access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties: { refreshToken: { type: string } }
      responses:
        '200':
          description: New token pair
          content: { application/json: { schema: { $ref: '#/components/schemas/TokenPair' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/logout:
    post:
      summary: Revoke a refresh token
      security: [{ bearerAuth: [] }]
      responses: { '204': { description: Logged out } }
  /api/v1/auth/me:
    get:
      summary: Current user profile
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Profile
          content: { application/json: { schema: { $ref: '#/components/schemas/UserResponse' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/mfa/totp/enroll:
    post:
      summary: Begin TOTP enrollment (returns provisioning URI)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Enrollment payload
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret: { type: string }
                  otpAuthUri: { type: string }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 12, writeOnly: true }
        firstName: { type: string }
        lastName: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, writeOnly: true }
        totpCode: { type: string, description: 'Required when MFA is enabled' }
    TokenPair:
      type: object
      required: [accessToken, refreshToken, tokenType, expiresIn]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        tokenType: { type: string, enum: [Bearer] }
        expiresIn: { type: integer, description: 'access token TTL seconds' }
    UserResponse:
      type: object
      required: [id, email, firstName, lastName, roles, createdAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        firstName: { type: string }
        lastName: { type: string }
        roles:
          type: array
          items: { type: string, enum: [CUSTOMER, ADMIN, OPERATOR] }
        mfaEnabled: { type: boolean }
        createdAt: { type: string, format: date-time }
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        traceId: { type: string }
  responses:
    BadRequest:
      description: Bad Request
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Unauthorized:
      description: Unauthorized
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Conflict:
      description: Conflict
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }